Hello triangle, install guide, update deps, minor refactors.
I'd normally try to avoid mixing multiple things into one commit,
but a lot of this code has been sitting there for over a month,
and it wouldn't be worth my time to try to disentangle it.

Also, usually I clean up my code a bunch before comitting
even if I have a feature working, but I did not do that here
either, for similar reasons (plus a lack of patience).
I intend to clean it up and return to my old quality standards
for future commits.